cloud_function/requirement.txt
google-cloud-logging
functions-framework